requirements.txt
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
executing==1.2.0
gunicorn==23.0.0
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
jupyter_client==8.0.1
jupyter_core==5.1.5
matplotlib-inline==0.1.6
mysqlclient==2.2.5
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pywin32==305
pyzmq==25.0.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.2
tornado==6.2
traitlets==5.8.1
tzdata==2024.2
wcwidth==0.2.6
